move completion/stp to language-server
